requirements.txt
django
python-dateutil
gunicorn
django-widget-tweaks
whitenoise
google-cloud-storage
xhtml2pdf
openpyxl
django-apscheduler
tablib